Finalize daemon run workflow lifecycle
